+++ pajbot/web/routes/base/login.py
@@ -26,7 +26,9 @@ def init(app):
 
         params = {
             "client_id": app.bot_config["twitchapi"]["client_id"],
-            "redirect_uri": app.bot_config["twitchapi"]["redirect_uri"],
+            "redirect_uri": app.bot_config["twitchapi"].get(
+                "redirect_uri", f"https://{app.bot_config['web']['domain']}/login/authorized"
+            ),
             "response_type": "code",
             "scope": " ".join(scopes),
             "state": json.dumps(state),